dotfiles: Refactoring to organize a bit.
I moved the brew and cask files into a packages directory, setup.sh into
a bin directory and the actual dotfiles into a configs directory because
it seems cleaner and more organized.

There is still an issue when sourcing .zshrc from bash that I need to
figure out.  Switching shells terminates the script and that's
troublesome.

I'd still like to add support for sudo instead of having to enter the
password when prompted by Homebrew and by pip.